Move default intrinsic type kinds from lib/semantics to lib/common
