+++ lib/common/default-kinds.cc
@@ -1,4 +1,4 @@
-// Copyright (c) 2018, NVIDIA CORPORATION.  All rights reserved.
+// Copyright (c) 2018-2019, NVIDIA CORPORATION.  All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,9 +13,9 @@
 // limitations under the License.
 
 #include "default-kinds.h"
-#include "../common/idioms.h"
+#include "idioms.h"
 
-namespace Fortran::semantics {
+namespace Fortran::common {
 
 IntrinsicTypeDefaultKinds::IntrinsicTypeDefaultKinds() {
 #if __x86_64__
