+++ lib/semantics/CMakeLists.txt
@@ -1,4 +1,4 @@
-# Copyright (c) 2018, NVIDIA CORPORATION.  All rights reserved.
+# Copyright (c) 2018-2019, NVIDIA CORPORATION.  All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -18,7 +18,6 @@ add_library(FortranSemantics
   attr.cc
   canonicalize-do.cc
   check-do-concurrent.cc
-  default-kinds.cc
   expression.cc
   mod-file.cc
   resolve-labels.cc
